+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Enhancements
     * Orphan files are removed when starting the database.
     * Add `:limit` option to `Goblin.scan`
+    * add `Goblin.has_key?` and `Goblin.Tx.has_key?` functions
 
 ## v0.11.0 (2026-06-16)
 * Breaking changes